requirements.txt
tqdm
scipy
biopython
einops
matplotlib
mrcfile
pandas
fair-esm
pyhmmer
loguru